Stop overwriting v3.Cluster on update
If the github.com/rancher/rancher/pkg/apis import is out of date and
rancher-operator attempts to update a v3.Cluster object, then any new
fields added to the cluster object in Rancher are deleted when
rancher-operator tries to update the cluster.

In order to prevent this behavior, Daishan added a patch call instead.
This change moves that code to a more general location to be used in
other places where the v3.Cluster is updated.